# priv/static/swagger.json
{
  "definitions": {
    "User": {
      "description": "A user in the application",
      "example": {
        "address": "Bandra Complex, Kurla Road",
        "city": "Mumbai",
        "date_of_birth": "2003/08/06",
        "district": "Mumbai",
        "email": "rahul.sharma@example.com",
        "full_name": "Rahul Sharma",
        "gender": "Male",
        "phone": "9998887777",
        "pincode": "400011",
        "role": "student",
        "state": "Maharashtra",
        "whatsapp_phone": "9998887777"
      },
      "properties": {
        "address": {
          "description": "Address",
          "type": "string"
        },
        "city": {
          "description": "City",
          "type": "string"
        },
        "date_of_birth": {
          "description": "Date of Birth",
          "type": "date"
        },
        "district": {
          "description": "District",
          "type": "string"
        },
        "email": {
          "description": "Email",
          "type": "string"
        },
        "full_name": {
          "description": "Full name",
          "type": "string"
        },
        "gender": {
          "description": "Gender",
          "type": "string"
        },
        "phone": {
          "description": "Phone number",
          "type": "string"
        },
        "pincode": {
          "description": "Pin code",
          "type": "string"
        },
        "role": {
          "description": "User role",
          "type": "string"
        },
        "state": {
          "description": "State",
          "type": "string"
        },
        "whatsapp_phone": {
          "description": "Whatsapp phone",
          "type": "string"
        }
      },
      "title": "User",
      "type": "object"
    },
...
    "Users": {
      "description": "All users in the application",
      "items": {
        "$ref": "#/definitions/User"
      },
      "title": "Users",
      "type": "array"
    },
...
    "UserSession": {
      "description": "A mapping between user and sesssion-occurence",
      "example": {
        "data": {
          "substitute-teacher-name": "Ms. Poonam"
        },
        "end_time": "2022-02-02T11:30:00Z",
        "is_user_valid": true,
        "session_occurrence_id": 2,
        "start_time": "2022-02-02T11:00:00Z",
        "user_id": "12213221211"
      },
      "properties": {
        "data": {
          "description": "Additional data for user session",
          "type": "map"
        },
        "end_time": {
          "description": "User session end time",
          "type": "timestamp"
        },
        "is_user_valid": {
          "description": "Signifies whether the user exist or not",
          "type": "boolean"
        },
        "session_occurrence_id": {
          "description": "The id of the session occurrence",
          "type": "integer"
        },
        "start_time": {
          "description": "User session start time",
          "type": "timestamp"
        },
        "user_id": {
          "description": "The id of the user",
          "type": "string"
        }
      },
      "title": "UserSession",
      "type": "object"
    },
...
    "UserSessions": {
      "description": "All user and session occurence mappings",
      "items": {
        "$ref": "#/definitions/UserSession"
      },
      "title": "UserSessions",
      "type": "array"
    },
    "SessionOccurrenceWithUser": {
      "description": "A single session occurrence with user details",
      "example": {
        "end_time": "2022-02-02T11:30:00Z",
        "session_id": 1,
        "start_time": "2022-02-02T11:00:00Z"
      },
      "properties": {
        "end_time": {
          "description": "Session occurrence finish time",
          "type": "timestamp"
        },
        "session_id": {
          "description": "Session ID",
          "type": "integer"
        },
        "start_time": {
          "description": "Session occurrence start time",
          "type": "timestamp"
        }
      },
      "title": "SessionOccurrenceWithUser",
      "type": "object"
    },
...
    "GroupIds": {
      "example": {
        "group_ids": [
          1,
          2
        ]
      },
      "properties": {
        "group_ids": {
          "description": "List of group ids",
          "type": "array"
        }
      },
      "type": "object"
    },
...
    "SessionOccurrences": {
      "description": "All the session occurrences ",
      "items": {
        "$ref": "#/definitions/SessionOccurrence"
      },
      "title": "SessionOccurrences",
      "type": "array"
    },
...
    "SessionOccurrence": {
      "description": "A session occurrence for a session",
      "example": {
        "end_time": "2022-02-02T11:30:00Z",
        "session_id": 1,
        "start_time": "2022-02-02T11:00:00Z"
      },
      "properties": {
        "end_time": {
          "description": "Session occurrence finish time",
          "type": "timestamp"
        },
        "session_id": {
          "description": "Session ID",
          "type": "integer"
        },
        "start_time": {
          "description": "Session occurrence start time",
          "type": "timestamp"
        }
      },
      "title": "SessionOccurrence",
      "type": "object"
    },
...
    "Session": {
      "description": "A session in the application",
      "example": {
        "created_by_id": 1,
        "end_time": "2022-02-02T11:30:00Z",
        "form_schema_id": 1,
        "is_active": true,
        "meta_data": {
          "substitute-teacher-name": "Ms. Poonam"
        },
        "name": "Kendriya Vidyalaya - Weekly Maths class 7",
        "owner_id": 2,
        "platform": "meet",
        "platform_link": "https://meet.google.com/asl-skas-qwe",
        "portal_link": "https://links.af.org/kv-wmc7",
        "repeat_schedule": {
          "repeat-till-date": "2022-12-31 11:59:59",
          "repeat-type": "weekly"
        },
        "session_id": "c714-e1d4-5a42-0f9f-36b3",
        "start_time": "2022-02-02T11:00:00Z"
      },
      "properties": {
        "created_by_id": {
          "description": "User ID for the session creator",
          "type": "integer"
        },
        "end_time": {
          "description": "Session finish time",
          "type": "timestamp"
        },
        "form_schema_id": {
          "description": "Id for the form schema",
          "type": "string"
        },
        "is_active": {
          "description": "Tells whether session is active or not",
          "type": "boolean"
        },
        "meta_data": {
          "description": "Additional meta data for the session",
          "type": "map"
        },
        "name": {
          "description": "First name",
          "type": "string"
        },
        "owner_id": {
          "description": "User ID for the session owner",
          "type": "integer"
        },
        "platform": {
          "description": "Platform where session being hosted",
          "type": "string"
        },
        "platform_link": {
          "description": "Link for the platform",
          "type": "string"
        },
        "portal_link": {
          "description": "Link generated by the portal",
          "type": "text"
        },
        "repeat_schedule": {
          "description": "Repeat type and repeat till date for session",
          "type": "json"
        },
        "session_id": {
          "description": "Id for the session",
          "type": "string"
        },
        "start_time": {
          "description": "Session start time",
          "type": "timestamp"
        }
      },
      "title": "Session",
      "type": "object"
    }
  },
  "info": {
    "title": "DB Service application",
    "version": "1.0"
  },
  "paths": {
...
    "/api/user": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.UserController.index",
        "parameters": [
          {
            "description": "The email of the user",
            "in": "query",
            "name": "email",
            "required": false,
            "type": "string"
          },
          {
            "description": "The full name of the user",
            "in": "query",
            "name": "full_name",
            "required": false,
            "type": "string"
          },
          {
            "description": "The date of birth of the user",
            "in": "query",
            "name": "date_of_birth",
            "required": false,
            "type": "string"
          },
          {
            "description": "The phone of the user",
            "in": "query",
            "name": "phone",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        },
        "summary": "",
        "tags": [
          "User"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.UserController.create",
        "parameters": [
          {
            "description": "User to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "summary": "",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/session/{id}/update-groups": {
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.SessionController.update_groups",
        "parameters": [
          {
            "description": "The id of the session record",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "List of group ids to update for the session",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupIds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Session"
            }
          }
        },
        "summary": "",
        "tags": [
          "Session"
        ]
      }
    },
    "/api/session-occurrence": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.SessionOccurenceController.index",
        "parameters": [
          {
            "description": "The id the session",
            "in": "query",
            "name": "session_id",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SessionOccurrences"
            }
          }
        },
        "summary": "",
        "tags": [
          "SessionOccurence"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.SessionOccurenceController.create",
        "parameters": [
          {
            "description": "Session occurence to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SessionOccurrence"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/SessionOccurrence"
            }
          }
        },
        "summary": "",
        "tags": [
          "SessionOccurence"
        ]
      }
    },
...
    "/api/user/{userId}": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.UserController.show",
        "parameters": [
          {
            "description": "The id of the user record",
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "summary": "",
        "tags": [
          "User"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.UserController.update",
        "parameters": [
          {
            "description": "The id of the user record",
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "User to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "summary": "",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/user-session": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.UserSessionController.index",
        "parameters": [
          {
            "description": "The id the user",
            "in": "query",
            "name": "user_id",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserSessions"
            }
          }
        },
        "summary": "",
        "tags": [
          "UserSession"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.UserSessionController.create",
        "parameters": [
          {
            "description": "User session to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserSession"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/UserSession"
            }
          }
        },
        "summary": "",
        "tags": [
          "UserSession"
        ]
      }
    },
...
    "/api/session-occurrence/{sessionOccurrenceId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.SessionOccurenceController.delete",
        "parameters": [
          {
            "description": "The id of the session occurence record",
            "in": "path",
            "name": "sessionOccurrenceId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "SessionOccurence"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.SessionOccurenceController.show",
        "parameters": [
          {
            "description": "The id of the session occurence record",
            "in": "path",
            "name": "sessionOccurrenceId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SessionOccurrenceWithUser"
            }
          }
        },
        "summary": "",
        "tags": [
          "SessionOccurence"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.SessionOccurenceController.update",
        "parameters": [
          {
            "description": "The id of the session occurrence",
            "in": "path",
            "name": "sessionOccurrenceId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Session occurrence to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SessionOccurrence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/SessionOccurrence"
            }
          }
        },
        "summary": "",
        "tags": [
          "SessionOccurence"
        ]
      }
    },
...
    "/api/user-session/{userSessionId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.UserSessionController.delete",
        "parameters": [
          {
            "description": "The id of the user session record",
            "in": "path",
            "name": "userSessionId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "UserSession"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.UserSessionController.show",
        "parameters": [
          {
            "description": "The id of the user session record",
            "in": "path",
            "name": "userSessionId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserSession"
            }
          }
        },
        "summary": "",
        "tags": [
          "UserSession"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.UserSessionController.update",
        "parameters": [
          {
            "description": "The id of the session record",
            "in": "path",
            "name": "userSessionId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "User session to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserSession"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/UserSession"
            }
          }
        },
        "summary": "",
        "tags": [
          "UserSession"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}